fix(autofix): Fix Github diff truncated files list when embedding for autofix (#138)
### Problem
Turns out the github compare API [returns a maximum of 300 files ](https://stackoverflow.com/a/71447308), breaking our code to embed the diff between the cached commit and the commit in the autofix request.

### Fix
Use the git diff API endpoint as the above stackoverflow post suggests which appears to return all file diffs (confirmed with a 1k file++ test run). We have to extract the git authorization headers from the PyGithub library and parse the diff locally. We also fix the issue with not correctly handling removed files.