fix(autofix): Allow stacktrace line_no to be None (#324)
Fixes [TIMESERIES-ANALYSIS-SERVICE-30](https://sentry.sentry.io/issues/5033928009/)

To address the ValidationError for StacktraceFrame during the creation, adjustments will be made to allow 'line_no' and 'col_no' to optionally be None. The string representation (to_str) of StacktraceFrame will be modified accordingly to handle the cases where 'line_no' and/or 'col_no' are None, ensuring clear notation for unknown lines or columns.