fix(autofix): Fix replacements introducing newlines (#419)
Tried on 2 issues locally, no more unexpected newlines introduced:

https://github.com/getsentry/seer/pull/418/files

https://github.com/getsentry/seer/pull/416/files

Issue was caused by the LLM getting confused with the extra newline in the XML tag contents, so it often added one. This fix ensures that there is no newline padding in the input to the LLM and only adds the newline after and ensures both the reference and replacement snippet end in a newline.