feat(codebase): Status endpoint for codebase index (#654)
Introduces the endpoint `/v1/automation/codebase/index/status` to get the current status of the codebase. Returns the following statuses:
```py
class CodebaseIndexStatus(str, enum.Enum):
    UP_TO_DATE = "up_to_date"
    INDEXING = "indexing"
    OUT_OF_DATE = "out_of_date"
    NOT_INDEXED = "not_indexed"
```

Now a namespace has the field `status` which can be 
```py
class CodebaseNamespaceStatus(str, enum.Enum):
    CREATED = "created"
    PENDING = "pending"
    UPDATING = "updating"
```
to do this, we change the initial codebase index creation step to create the repo and namespace db rows before the indexing is complete. Because of this in the case of a failure during embedding etc it will need to delete the namespace