ref(codebase): Split codebase create and indexing (#663)
Splits out the codebase index creation into a separate creation of the db entries and the index/embedding step. This was needed to be done to return the indexing status to Sentry immediately. In the previous version, it would've needed to wait for a celery worker to pick up the job before the indexing status is present.

This actually plays nicely into splitting up logic that doesn't need the GPU power out (creating the db entries) from the work that does, which is sent to the celery worker.