feat(autofix): Step-based pipeline architecture (#687)
Introduces the concept of a step architecture based around celery tasks to Autofix.

Each `PipelineStep` is a single piece of work scheduled as a celery task.
Each step can be "baked" into a [celery signature](https://docs.celeryq.dev/en/stable/userguide/canvas.html#signatures) by calling `get_signature` on it. Arguments and any options will be baked with it as well. This means we can bake the `queue` option right into the signature.
These signatures can be scheduled to run on a celery queue and also serialized and deserialized. This allows us to define _chains_ of steps and conditional running of steps.

Although it's based on celery for now, if/when we decide to move away from it to @corps's solution, this should be set up nicely for that transition given we've separated out the pieces of work.

#### Steps that have been implemented for autofix
- Create missing indexes (CPU)
- Parallel codebase indexing (GPU)
- Root Cause (CPU)
- Execution (GPU)

#### Next Steps
- Split out the execution step out so it doesn't have to entirely hog the GPU queue.