feat(codebase): Check files match up with repo & handle accordingly (#706)
Again, ideally we don't have missing files or corrupt codebase indexes, but again to echo @corps, _a good system anticipates failures_ so now we will have one which will automatically fix these issues or in what I think are rarer cases will log a message to notify us.

We check the integrity of a codebase index by comparing the list of files with the list of files in the index, and make sure they match. If not, it will recreate the codebase index.

#### Expectations
We might see this recreate often early on as we've had problems with messed up indexes earlier during internal usage, but ideally once those are fixed they shouldn't happen again. If they do then there must be some incorrect logic in the codebase updating logic.