feat(autofix): Local keyword search tool (#846)
Adds a keyword search tool that the LLM can use.

The keyword search tool:
1. Downloads the repo contents to a temp directory
2. Runs a keyword search that matches the file contents and file path
3. (Optional) If a starting path is provided, results are ranked
according to proximity to that path
4. Deletes the temp directory

<img width="1365" alt="Screenshot 2024-07-04 at 12 05 47 AM"
src="https://github.com/getsentry/seer/assets/30991498/180c44e4-8521-4f9d-bc2e-b8ff4b9423df">

**Note**: It's pretty inefficient to download and delete the repo
everytime this tool is used, we should be downloading it for a step and
clearing the tmpdir at the end of the step run. However, as this is for
the experimental codebase-indexless run, we should go with this for now
to reduce code changes and touch less files. As we're planning on doing
a big refactor once that decision is made.

If we go ahead with it we should refactor the downloading to work as
described above.