(autofix) Add Claude agent (#927)
Adds a Claude client and agent as an alternative to ChatGPT for running
Autofix (issue https://github.com/getsentry/seer/issues/900). Also
includes refactors that make the code more extensible to support
multiple models.

Note: there is currently no way of switching to Claude, besides going
into the necessary files and switching GptAgent or GptClient to
ClaudeAgent or ClaudeClient. This is a future task, depending on how we
decide to leverage these models.

---------

Co-authored-by: jenn.muengtaweepongsa <jenn.muengtaweepongsa@sentry.io>
Co-authored-by: Jenn Mueng <30991498+jennmueng@users.noreply.github.com>