(autofix) Add simple PII scrubber for stack traces and variables sent to LLMs (#942)
Use regex to redact basic PII like emails, SSNs, and more, before stack
traces and runtime variables are sent to an LLM. We saw this occurring
in some Langfuse logs and it would be a good privacy safeguard for our
sensitive customers.

Of course, it's just regex, so it's not perfect. It may miss some things
and it may scrub data that's formatted like PII, but isn't. Some regex
rules are also less effective in non-US and non-English-language
contexts.

However, the method added can easily be extended with more/better rules
in the future.

Issue #941