feat(autofix): Run full evals from langfuse dataset within seer (#873)
Introduces the ability to run full autofix evaluation runs on a langfuse
dataset within seer. Just hit `/v1/automation/autofix/evaluations/start`
with the name of the dataset and a name to name your run. It will
schedule the evaluation runs onto the celery CPU queue.

This is mainly to run in a dev environment pointed towards a dev
langfuse instance. However, once merged and deployed this also means we
can run evals within the production environment against production
langfuse.