ref(autofix): move all sha stuff into the repo client (#868)
`RepoClient` will default to using the default branch's head commit sha
unless a base commit sha is passed in the repo definition. The only time
(for now) that this will be done is in evals.

The `RepoDefinition` model now stores a `base_commit_sha` field for the
respective repo. The `base_commit_sha` field on `AutofixRequest` is now
removed as it no longer makes sense as we support multiple repositories
now.