feat(autofix): Root cause evals (#896)
Introduce evals for just the root cause pipeline.

Will output 4 scores for a run item:
- `"highest_score_gpt4o_5x_panel"`: The score for the highest scored
cause out of all the returned root causes.
- `"positioning_score_gpt4o_5x_panel"`: Positioning of the highest
scored cause, if the highest scored cause is first, this score is `1.0`.
If it is last, it will be `0.0`. The score is calculated proportional to
the number of causes provided.
- `"mean_score_gpt4o_5x_panel"`: The mean score of all the root causes.
- `"error_weighted_score_gpt4o_5x_panel"` This score is the same as the
highest score but scored 0 if there is an error or no root cause
returned. This is used to weight the score in the aggregated run result.

Run a root cause eval by setting
```
"run_type: "root_cause"
```
in the POST request body to the evaluation start endpoint