chore(auth): Exclude URL from auth check as it can differ between environments (#1142)
By using URL as part of the payload that is signed we are assuming that
URL used by client will be same as the URL seen in the request payload
on Seer end. This is not true as requests from Subscription processor
has full UTL including server name while those from the UI do not.

This PR makes URL optional. This will be step 1 of the fix. We need to
follow this up with two more changes:
* remove URL from all locations on the client end where we are signing
Seer requests
* once we confirm that Seer is not getting requests with URL anymore,
then remove the code that checks for URL + payload.

---------

Co-authored-by: Zachary Collins <zachary.collins@sentry.io>