feat(autofix): Add controllable insights chain (#1213)
When an certain update request is received, we jump back to the selected
point in the steps/insights chain, delete any steps, insights, and
memory after that, add the user feedback to the memory, and continue
analyzing.

This allows the user full control over the model's reasoning, and let's
them rethink and guide the model at any point. This also allows
restarting from errored/no-root cause states, and jumping back to
previous steps.