feat(autofix): Support user editing code changes (#1308)
Allows the user to edit or reject hunks of the code changes. We update
the changes stored in the step.

This also required changing the PR creation logic to support creating a
PR from these updated diffs (`FilePatch`) as well as the old method
(`FileChange`).

One limitation is that there is no unique commit message for each file
patch.