chore(anomaly_detection): apply direction logic for alerts with both as direction (#1366)
This change ensures that the direction logic is applied for anomalous
points even when the alert is configured for both directions. This
change helps reduce noise from anomalies that do not breach any
threshold.

Note: Apart the core logic of enabling direction logic, this PR also
includes multiple fixes for the time stamps not being handled correctly
that surfaced due to this change.