fix(autofix): Bump updated_at BEFORE throwing the state in the dbmodel (#1575)
Runs were timing out because there actually are two `updated_at` fields,
the main one being used in the class instances are the ones inside the
`value` json. We're supposed to be syncing the field on the db model
with the one from the json for the sake of db querying, but there was a
bug in the big PR that updates this field in the value AFTER the value
json was set.

So the updated_at inside the json was the original one, never bumped.