feat(autofix): Stream output in Autofix agent (#1617)
Replaces Autofix agent LLM calls with streaming. Saves the output as it
comes in the step object in the run state. All other functionality is
identical.

I added a function for streaming with OpenAI as well, and tested that it
works, but we don't use it at the moment. Just future proofing.

---------

Co-authored-by: Jenn Mueng <30991498+jennmueng@users.noreply.github.com>