chore(anomaly detection): reduce prophet inference in batch mode to improve performance (#1700)
This change limits the number of prophet inferences in batch detection
mode to keep the store_data end point bounded. In this iteration we are
limiting the location detection logic to just last 12 data points in the
batch mode. In a follow-up we will also switch to invoking prophet just
once for those 12 data points.