fix(autofix): Remove langfuse flushes to reduce cold starts (#1694)
Langfuse automatically flushes in the background. Correct me if I'm
wrong, but I don't think we ever force terminate the program and would
lose any data by relying on the automatic flushes.

These manual flushes block the main program and add ~750ms-1s of cold
start on every LLM call, every keyword search, and every missing file
fixer. (some profiles showed it taking up to 10s sometimes) Langfuse
officially says to only use this when necessary as it is known to hurt
performance.

So this would be a nice improvement to snappiness if we can get away
with it.