fix(api): Reintroduce request signing (#1762)
_Previously request signing was disabled due to a quirk in our
networking setup that made the request url passed to the client signer
not the same as the request seen on the server side._

This reimplements the request signing with an optional nonce but without
the salted url verification. This is a stop-gap measure to reimplement
signature signing for now until we have bandwidth to invest in a more
complete solution.