feat(autofix): Add unit test options to coding step (#1880)
The coding step can be invoked with options to just write a fix, to just
write a repro test, or both. It defaults to just a fix.
Evals are stable between writing a fix and writing a fix with a test.

This PR should go in before the frontend change.