fix(autofix): Some fixes (#1904)
Not going ahead with DeepClaude, but discovered some bug in converting
tool calls between OpenAI, Anthropic, and Gemini, so fixing that. Also
adding support for reasoning-related parameters for OpenAI.

Also improving the get_valid_file_paths check that the AI suggested.

And fixes the exception logging in get_file_content so we don't get
spammed by Sentry.

And change the index at which we insert the initial prompt in the
solution/coding step to prevent some hallucinations we were seeing.