fix(autofix): Add fallback for code application step (#1938)
Occasionally, the apply model would not complete its response with a
`</updated_code>` tag or it would be truncated for some other reason,
causing the "new content" to be empty and cause the whole file to be
deleted.

Now if there is no new content from the apply model (meaning something
went wrong), fall back to asking Gemini Flash to directly write a
unified diff. (I tested always just writing a diff, but the fix accuracy
went down, so just using it as a fallback).

---------

Co-authored-by: Jenn Mueng <30991498+jennmueng@users.noreply.github.com>