chore(evals): Improve final fix scoring prompt (#1963)
An issue with the prior prompt was that we were having it overfocus on
scoring how well the predicted solution "matches" an expected output,
rather than scoring how well the issue is fixed.

This prompt update:
1. Changes the concept of an "expected" solution to a "known" solution
and prompts o3-mini to use it to inform what the issue was
2. The score is now on how well the predicted solution fixes the issue
rather than asking it to score how well it "matches"
3. Prompt format adjustment to place context at the end for reasoning
models.

(ignore the latency, i was running multiple evals at the same time so it
slowed)
![CleanShot 2025-02-19 at 15 22
39@2x](https://github.com/user-attachments/assets/c042422c-5bb4-4a4b-b134-ef40e09b8261)