feat(autofix-evals): Separate solution & coding evals (#2020)
The solution eval now tests the fix on the solution step, while the
coding step we have a new `correctness_score` and a `conciseness_score`
that just tests how well it takes its instructions and writes code for
it.

This also removes the `run_type` field from the evaluation call and the
individual root cause and coding evals, this is because we weren't using
the other ones. They were outdated deadcode. We can just quickly write
them up again when we were using it.