fix(autofix): Better PR title + error handling for long PR creation (#2041)
Now that we know the GitHub API takes a hot second we can design around
this.
From looking at traces in Sentry, the gateway times out at 45s

1. Put the commit changes in a celery task with timeouts that we wait
for, 25s soft timeout & 30s hard timeout
2. We make it so the branch creation isn't done again if it already
exists.
3. The update endpoint will now return a message, and an error status
with a message that can be surfaced. Following up with a tiny Sentry
side backend PR to forward the status and message, and a tiny frontend
PR to actually use it

This also updates the PR title to be lowercase and prefixed with `fix:`
and the branch names should be shorter, more targeted as the Change
Describer generates them now.

---------

Co-authored-by: Rohan Agarwal <47861399+roaga@users.noreply.github.com>