feat(autofix): Generate summaries for RC and Solution timelines (#2071)
Has the formatter step for Root Cause and Summary return a 1 liner
summary of the root cause and solution. Places it in an optional
`description` field for both steps, and should be backwards compatible.

Does not meaningfully affect eval scores.
Root Cause stable, but helpfulness 0.8969 -> 0.9196
Solution seems stable but is fixed 0.5670 -> 0.5804, score 0.6524 ->
0.6454
Coding dropped, even tho it wasn't touched? conciseness 0.9622 ->
0.9396, correctness 0.9304 -> 0.8983