feat(autofix): Add confidence scoring to autofix outputs (#2081)
At the end of every main Autofix step, asks an LLM with context of the
full agent memory to reflect on uncertainties, assumptions, open
questions, etc. Returns a confidence score on the correctness of the
output and a score of confidence in achieving the next step (e.g.
confidence in root cause + confidence in being able to figure out a
solution).

Also optionally returns a comment raising uncertainties or questions to
the user. This comment thread can be interacted with in the same way as
the existing user-initiated threads. Replaces the Q&A tool.

I haven't done a formal evaluation of the confidence scores but they
seem reasonable anecdotally.

We can use them later when we do auto-triggers to determine how far in
the run to go.