fix(autofix): Stricter fuzzy matcher for more reliable diff application (#2099)
This hopefully fixes the issue where the diff application looks
completely glitched. My hypothesis from the examples I looked at was
that it was replacing snippets in the wrong location, so a stricter
matching process should fix it. This is also ok because our diff output
from the coding step is nearly perfect in accuracy, so we don't need to
be as "fuzzy" as we used to.