feat(anomaly detection): MP Override (#2105)
- Adjusts the scoring algorithm to use the MP Flags if there is strong
confidence (mp_dist > 2 * threshold) even if prophet does not indicate
an anomaly. This is an incremental change.

Before MP Override:
![Screenshot 2025-03-05 at 9 31
41 PM](https://github.com/user-attachments/assets/768c1322-e051-4280-b416-6fa9cac8cd69)

After MP Override:
<img width="1414" alt="Screenshot 2025-03-05 at 9 54 36 PM"
src="https://github.com/user-attachments/assets/185583b2-cbab-4a5f-bd35-5889c45172f9"
/>

If we look at the end of the big spike in this example, because the
Prophet CI begins to widen after, originally we stop the alert earlier
than we should (not good in stream case where we are not sure how the
timeseries will behave next). By adding the override, we extend the
anomaly just a little more and ensure we are fully confident before
marking it as resolved.

Overall in the other examples, only 1 of them produced a 1 timestep
false positive while the rest remained the same as intended.

**CAVEAT:**
An example like this still needs to be dug deeper into because this
spike at Nov 15th should be called out, but the matrix profile does not
go high enough for it to trigger and Prophet scores are not high enough
for it to override either due to the confidence intervals being so
large. On `high` sensitivity though this should be captured due to
confidence intervals shrinking.

![image](https://github.com/user-attachments/assets/ccdbc44c-1dab-4cb7-9fd8-e15fbc9ac574)