feat(autofix): Receive trace tree in request payload (#2144)
Creates a model for trace data for the `sentry` backend to send. Also
passes it into the root cause request, but doesn't use it, so I can
inspect Langfuse logs to see the data looks right in prod before putting
it in a prompt. It does include the transaction name in the event
details that go in our prompts though.

Once I can confirm this is working in prod, my plan is to follow up with
using this in our prompts and adding related tools.