fix(autofix): Return existing PR + logging (#2175)
After looking at the trace of a lot of the errors, [a lot of them are
coming from the non-draft pr
line](https://sentry.sentry.io/issues/6361334163/events/f3e7f38368fa449aba774b4fa940d9fa/?project=6178942&sort=date&statsPeriod=14d).
And apart from that the errors don't have a stacktrace, this attempts to
better logging there so we can debug what the actual issue is.

Also, if a PR already exists for a branch, we should just return it and
link to it on Autofix cuz u can only have 1 PR for a branch...

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>