feat(autofix): Preferences for Autofix w/ repo & branch selection (#2248)
Introduces a `SeerPreferences` which will store the preferences on a
per-project basis and can be used for all seer uses not just Autofix.

Has Autofix use the instructions in the prompt, and the branch in the
branch/PR creation.
- If you provide an invalid branch, Autofix will just error when you try
to run, this is a first pass and in the near future we can follow up
with an actual check during settings edit whether the branch is valid or
not
![CleanShot 2025-03-24 at 23 51
57@2x](https://github.com/user-attachments/assets/9260c2aa-3e91-4219-9490-6b5c9e0d6be3)

- You'll need to start a new run to apply the changes, we would need to
redo a lot of the run memory logic to allow for hot-swapping out user
messages with repo/instruction information
![CleanShot 2025-03-24 at 23 26
25@2x](https://github.com/user-attachments/assets/5172d972-6038-47c5-9c16-88e800a7d5f5)



...the code codecov is complaining about missing coverage I'm pretty
sure is covered lol

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>
Co-authored-by: Rohan Agarwal <47861399+roaga@users.noreply.github.com>