fix(autofix): Handle truncated repos (#2326)
When getting a git tree, if the result from the GitHub API is truncated,
recursively get the subtrees of the main tree's directories until we
have a complete tree. Since this is done in parallel, it's pretty fast.
On Seer, a non-truncated call is ~1.2s and this recursive approach is
~3s. On getsentry, it was ~1s and ~2.5s.