feat(autofix): Use claude tools in coding (#2309)
Introduces the new claude coding tools for use by the coding agent.

Some tools such as the `view` tool's path viewing and file viewing are
redundant but required for the claude tool to work. We're keeping both
implementations because ours is better and evals are doing good.

Also adds the edits programattically as insight cards + allows for
rethinks

93.65% correctness and 98.97% conciseness

![CleanShot 2025-03-31 at 15 23
48](https://github.com/user-attachments/assets/fe904496-bf60-40b5-9164-921d56266d3e)

NOTE: Frontend needs to go out before this

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>