[feat] Compare generated PR comments against embeddings before posting (#2329)
The logic is as follows

After generating PR review comments for a PR, for each comment, we check
if it is a `positive_comment`.

We do this by first fetching all similar comments for the PR's owner in
the past(defined by cosine distance and a similarity threshold). If
there are less than 5 similar comments, we will optimistically approve
the comment for posting.

If there are 5 similar comments, we will only post the generated comment
if the majority of the similar comments are `positive`(more upvotes than
downvotes)

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>
Co-authored-by: codecov-ai[bot] <156709835+codecov-ai[bot]@users.noreply.github.com>