chore(relevant_warnings): data handling
* Makes sure we return the new suggestions in a format that Overwatch understands

Because we want minimal changes to the app at this point we will transform the new suggestion
style into something compatible to what overwatch is currently expecting.
We are losing a bit of context from the suggestions, but it shouldn't matter much at this point.
This is _supposed_ to be a temporary thing

* Make sure the LLM has access to issue and warnings ID

Overwatch expects warnings and issues IDs, and because we just shove everything at the LLM at once,
we kinda expect it to return the IDs for us (and hopefully not hallucinate new IDs).
For that the LLM needs visibility on the IDs, so changing formatting of issues and warnigns a bit
to include the IDs.