change relevant warnings prediction (#2398)
Significant changes to the Relevant Warnings Step so that we suggestions
are better.

* Don't require warning _and_ issue to exist
* Add diff to the analysis
* Change prompt to _predict_ potential issues

This is a significant change to the relevant warnings step. It moves
somewhat away from the warnings as the sole main focus, and kinda
equally use warnings and issues as extra context for the LLM to evaluate
the PR dif.

It changes several smaller LLM calls (that decide if warnings and issues
can be paired logically) by 1 bigger one (that analyses the entire
diff). Works well for small diffs and a small number of warnings and
issues, but we're not very sure the scaling factor here.

The next steps would be to better limit the size of the diff, or break
it up into different chunks and analyze the chunks individually.