fix(autofix-evals): move score outside observation (#2412)
Moves the scoring calls outside the main eval item observation context
manager. This is part of fixing the bug we saw where we had nested eval
runs.

The 2nd part of that is we **NEED** to run each eval run with a unique
name, even if you deleted the previous one