fix(autofix): Parallelize grep search (#2454)
If we run a bad grep query that times out when multiple repos are
available, it seems like the run freezes (and often it'll time out).
e.g. 45 second timeout * 4 selected repos = 3 min of doing nothing

So this PR parallelizes grep calls if there are multiple repos so that
at most we wait only 45 seconds.