fix(autofix): Fix using commit sha in tree sha (#2493)
We were supposed to be passing a `SHA1` type [tree object
sha](https://docs.github.com/en/rest/git/trees?apiVersion=2022-11-28#get-a-tree)
to the tree endpoint not the commit sha. I'm not sure how it worked
before when we were passing the commit sha though, maybe they were
unraveling it for us as a convenience/fallback?

ChatGPT o3 is saying GitHub rolled back the unravelling earlier this
year but didn't provide a source, could be hallucinating.

Regardless, we should fix this bug.