feat(autofix): Use much faster ripgrep in place of grep (#2508)
ripgrep should be much faster to run vs grep. We also rework the tool to
be more defined with specific parameters rather than letting the LLM
loose with grep.

Hard to get a real apples to apples comparison of speed from our evals
given that I'd have to spin up a notebook to pull langfuse spans, so
next best thing is a comparison:

Example of similar grep vs a much faster ripgrep:

![CleanShot 2025-05-01 at 02 32
26@2x](https://github.com/user-attachments/assets/8a844bab-1d7e-4f3b-b164-ecd063984fbd)

![CleanShot 2025-05-01 at 02 24
31@2x](https://github.com/user-attachments/assets/f8517323-c919-4be5-8a83-1bec30d3af26)

After using it a couple of times, in a couple eval runs, seems to be no
real memory issues caused by ripgrep itself

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>