fix: adjust item format
at the same time I was building the `dataset.py` other members of the team created their own script
to upload items to Langfuse. Obviously it generated 2 standards and that's bad.

So I'm adjusting this to match that item's format (so we can process those)

Classic case of lack of communication leading to extra work.