fix evaluation for relevant bugs
Change the evaluation model to be based on the WHAT / WHERE / HOW
as defined here: https://www.notion.so/sentry/Bug-prediction-1868b10e4b5d809895b3f09311cf50c4#1e58b10e4b5d804aab71f0c8d8da7171

Important to note, it's my interpretation of it.
The "WHERE" is not optional yet because the actual bugs and suggestions currently
need to be attached to some code location, so I'm using that to compare.

The comparison works by basically providing both sets (actual_bugs, suggested_bugs) to an LLM
and asking it to pair and score pairs.