fix(autofix): Fast fail expand_document when file doesn't exist (#2519)
**Main fix is:** We shouldn't be falling back to waiting & checking on
the downloaded repo for `expand_document` if the file path doesn't even
exist.

**Included refactors:**
This PR also changes the behavior so we move the file path fixing to the
tool call and outside of the Autofix context. This is because any stored
file path that's used elsewhere such as the memory prefill should be
made correct anyways, and isolates LLM-related feedback behavior to the
tool. We should be isolating any LLM-specific error handling and usage
to the tool set because we use the Autofix context elsewhere

`get_index_file_set` and `get_valid_file_paths` from `RepoClient` were
redundant, and the former was originally for when we were making
codebase indexes. Refactors all uses to use the latter.

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>