utility to see run results
Langfuse UI sucks and was making it hard for me to understand the actual
status of the system. No doubt because this is early runs and whatnot.

I probabl should have asked what we use for real, but I didn't, and just built this
real quick to get some better aggregated results.

So there you go