fix(autofix): Add None check for repo names in insight sharing (#2542)
The problem was that if the repo name from the insight was invalid (e.g.
"N/A" because the LLM knew there were no accessible repos), then
autocorrect would return None. Passing None into get_repo_client would
pick a non-accessible repo if that was the only one.

So we just make sure that we have a repo name before trying to get a
repo client.