feat(autofix): Use a git client to clone and checkout repos (#2544)
This PR changes how we download repos from downloading repo archives to
cloning the repo with `depth=1` and then checking out a specific commit
sha.

- This sets us up to be able to cache and load repos from GCS using this
`RepoManager`
- As the repo is downloading, we also update the Autofix state to ensure
that the state does not get timed out

## Implementation
Introduce the `RepoManager`, conceptually it's supposed to be the class
that manages a repo's download, caching etc. It should encapsulate the
entire lifecycle of a downloaded repo. It's also entirely source control
provider agnostic, and instead relies on getting passed a `RepoClient`
that it interacts with.

## How it works
1. The tools are initialized, the `RepoManager`s are created and then
called to initialize each on its own thread in the background.
2. The repo is cloned with `depth=1`
3. The repo is fetched at a specific commit sha from the `RepoClient`
and checked out at that specific sha.
4. Once a tool is called that needs to use the downloaded repo, via
`BaseTools._ensure_repos_downloaded`, we wait for the futures with a
`120s` timeout.
5. If a repo download times out, we tell the LLM that the repo is not
usable and to not use that tool with the repo.

## Next Steps
1. Setup GCS buckets in the required regions
2. Implement the GCS upload/download in the `RepoManager`

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>